one more chance

+++ mptestagent.py
@@ -114,7 +114,7 @@ class Player():
         inputs = layers.Input(input_shape)
         x = layers.Reshape((inputs.shape[1],
                             inputs.shape[2]*inputs.shape[3]))(inputs)
-        x = layers.Conv1D(512, 7, activation='relu')(x)
+        x = layers.Conv1D(512, 7, activation='relu', dtype='float32')(x)
         x = layers.Conv1D(512, 7,activation='relu')(x)
         x = layers.Conv1D(512, 7,activation='relu')(x)
         x = layers.Conv1D(512, 7,activation='relu')(x)
@@ -131,7 +131,7 @@ class Player():
 
     def brain_layers(self, x):
         x = layers.Flatten()(x)
-        x = layers.Dense(256, activation='relu')(x)
+        x = layers.Dense(256, activation='relu',dtype='float32')(x)
         x = layers.Dense(8192, activation='relu')(x)
         x = layers.Dense(1024, activation='relu')(x)
         x = layers.Dense(1024, activation='relu')(x)
